set base for e2e test selection rework
